requirements.txt
fire
pydot
GitPython
python-hcl2
graphviz
pytm
pyecore
pyyaml
semgrep